requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.23.2
pydantic==2.5.2
pyodbc==4.0.39
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
